improve CLI, help, factor out CLI and main to separate from functions
