+++ meta.yaml
@@ -1,9 +1,9 @@
 package:
   name: protoci
-  version: 0.0.0
+  version: 0.0.1
 
 source:
-  git_rev: 0.0.0
+  git_rev: 0.0.1
   git_url: https://github.com/ContinuumIO/ProtoCI
 
 build:
